+++ lib/240326/model/user.dart
@@ -1,37 +1,3 @@
-import 'dart:convert';
-
-import 'package:http/http.dart' as http;
-
-void main() async {
-  try {
-    final users = UserApi();
-    users.getUsers();
-  } catch (e) {
-    print(e);
-  }
-}
-
-class UserApi {
-  Null get jsonListString => null;
-
-  Future<List<User>> getUsers() async {
-    final http.Response response =
-        await http.get(Uri.parse('https://jsonplaceholder.typicode.com/users'));
-
-    if (response.statusCode == 200) {
-      List json = jsonDecode(response.body);
-
-      List<User> data = json.map((json) => User.fromJson(json)).toList();
-      return data;
-    } else {
-      throw Exception('Response 에러');
-    }
-
-    // Future<List<User>> users =
-    //     jsonList.map((e) => User.fromJson(jsonList)).toList();
-  }
-}
-
 class User {
   final int id;
   final String name;
